AHM-7 Prevent possible loss of precision in CodeQl check.

+++ Stefanini.ViaReport.Core/Business/DashboardBusiness.cs
@@ -83,7 +83,7 @@ namespace Stefanini.ViaReport.Core.Business
         }
 
         private static decimal CalculateAverage(long total, int count)
-            => total / count;
+            => (decimal)total / (decimal)count;
 
         private static DateTime? GetResolvedDate(IssueDto issue)
             => issue.Fields.Resolutiondate?.Date
